alpha 0.2 final code structure
This the structure of the processing system. The document processing
has been removed to a separate function. This should make it easier
to produce documentation of different shapes for different targets
easier. You will just need to run the processor with different parameters
after all the data collection has been done.